requirements.txt
aiohttp==3.7.4
alpaca-trade-api==1.5.1
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
deprecation==2.1.0
Flask==2.0.3
google-api-core==2.7.1
google-auth==2.6.0
google-cloud-secret-manager==2.9.2
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
msgpack==1.0.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
websocket-client==1.3.1
websockets==10.2
Werkzeug==2.0.3
yarl==1.7.2
